+++ homeassistant/components/tesla_fleet/config_flow.py
@@ -8,12 +8,11 @@ from typing import Any
 
 import jwt
 
-from homeassistant.components.application_credentials import ClientCredential
 from homeassistant.config_entries import ConfigEntry, ConfigFlowResult
 from homeassistant.helpers import config_entry_oauth2_flow
 
-from .application_credentials import TeslaOAuth2Implementation
-from .const import CLIENT_ID, DOMAIN, LOGGER, NAME
+from .const import DOMAIN, LOGGER
+from .oauth import TeslaSystemImplementation
 
 
 class OAuth2FlowHandler(
@@ -35,9 +34,7 @@ class OAuth2FlowHandler(
         """Handle a flow start."""
         self.async_register_implementation(
             self.hass,
-            TeslaOAuth2Implementation(
-                self.hass, DOMAIN, ClientCredential(CLIENT_ID, "", NAME)
-            ),
+            TeslaSystemImplementation(self.hass),
         )
 
         return await super().async_step_user()
